fix: use session based model and request storage
