+++ apis/nucleus/src/contexts/InstanceContext.js
@@ -7,4 +7,6 @@ export default React.createContext({
   constraints: {},
   interactions: {},
   themeApi: null,
+  modelStore: {},
+  selectionStore: {},
 });
