+++ apis/nucleus/src/hooks/useAppSelections.js
@@ -1,9 +1,10 @@
 /* eslint no-underscore-dangle: 0 */
-import { useEffect } from 'react';
-import { useAppSelectionsStore, modalObjectStore, appModalStore } from '../stores/selections-store';
+import { useContext, useEffect } from 'react';
+import InstanceContext from '../contexts/InstanceContext';
 
-function createAppSelections({ app }) {
+function createAppSelections({ app, store }) {
   const key = `${app.id}`;
+  const { modalObjectStore, appModalStore } = store;
 
   const end = async (accept = true) => {
     const { model, objectSelections } = modalObjectStore.get(key) || {};
@@ -80,18 +81,23 @@ function createAppSelections({ app }) {
   };
   return appSelections;
 }
-export default function useAppSelections(app) {
+export default function useAppSelections(app, selectionStore) {
   if (!app.session) {
     // assume the app is mocked if session is undefined
     return [];
   }
-  const [appSelectionsStore] = useAppSelectionsStore();
+  let store = selectionStore;
+  if (!store) {
+    store = useContext(InstanceContext).selectionStore;
+  }
+
+  const [appSelectionsStore] = store.useAppSelectionsStore();
   const key = app ? app.id : null;
   let appSelections = appSelectionsStore.get(key);
 
   useEffect(() => {
     if (!app || appSelections) return;
-    appSelections = createAppSelections({ app });
+    appSelections = createAppSelections({ app, store });
     appSelectionsStore.set(key, appSelections);
     appSelectionsStore.dispatch(true);
   }, [app]);
