+++ apis/nucleus/src/hooks/useSessionModel.js
@@ -1,8 +1,9 @@
 import { useEffect, useState } from 'react';
-import { useModelStore, useRpcRequestSessionModelStore } from '../stores/model-store';
+import { useModelStore, useRpcRequestSessionModelStore } from './useModelStore';
 
 export default function useSessionModel(definition, app, ...deps) {
   const key = app ? `${app.id}/${JSON.stringify(definition)}` : null;
+
   const [modelStore] = useModelStore();
   const [rpcRequestSessionModelStore] = useRpcRequestSessionModelStore();
   const [model, setModel] = useState();
