--- apis/nucleus/src/stores/selections-store.js
@@ -1,14 +1,0 @@
-import createKeyStore from './create-key-store';
-
-const [useAppSelectionsStore, appSelectionsStore] = createKeyStore({});
-const [useAppModalStore, appModalStore] = createKeyStore({});
-const [useModalObjectStore, modalObjectStore] = createKeyStore({});
-
-export {
-  useAppSelectionsStore,
-  useAppModalStore,
-  appSelectionsStore,
-  appModalStore,
-  useModalObjectStore,
-  modalObjectStore,
-};
